% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_utils.R
\name{add_simscore}
\alias{add_simscore}
\title{add similarity score of compounds}
\usage{
add_simscore(df)
}
\arguments{
\item{df}{A tibble result from \code{compound_similarity}.}
}
\description{
add similarity score of compounds
}
\keyword{internal}
