% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_cluster.R
\name{Addclusterscore}
\alias{Addclusterscore}
\title{Score and Rank Network Clusters}
\usage{
Addclusterscore(g, cluster_attr = "louvain_cluster", min_size = 3)
}
\arguments{
\item{g}{An \code{igraph} object. The graph must have a vertex attribute containing cluster labels.}

\item{cluster_attr}{Character. The name of the vertex attribute that stores cluster labels. Default is louvain_cluster.}

\item{min_size}{Integer. Clusters smaller than this size will be ignored. Default is 3.}
}
\value{
A data frame containing cluster statistics, ranked by Score.
}
\description{
This function evaluates the clusters/modules identified in the graph.
It calculates a score for each cluster based on density and size (MCODE style),
and returns a ranked data frame.
}
\examples{
data(demo_ppi)
ppi <- run_louvain(demo_ppi, resolution = 1)
louvain_score <- Addclusterscore(ppi, cluster_attr = "louvain_cluster", min_size = 3)
head(louvain_score)

}
