% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_utils.R
\name{compound_similarity}
\alias{compound_similarity}
\title{Similarity search (Tanimoto) on PubChem}
\usage{
compound_similarity(
  query,
  from = c("smiles", "cid", "inchikey", "name"),
  threshold = 90,
  topn = 10,
  fetch_factor = 3,
  compute_score = TRUE,
  ...
)
}
\arguments{
\item{query}{identifier (CID/SMILES/InChIKey/Name)}

\item{from}{one of c("cid","smiles","inchikey","name")}

\item{threshold}{integer 0–100 (percent), e.g. 90}

\item{topn}{max records to return. Default 10.}

\item{fetch_factor}{The multiplier for overfetching candidate compounds (default = 5).}

\item{compute_score}{Logical. If TRUE, use rcdk to compute Tanimoto score locally. Set to FALSE if rJava causes crashes.}

\item{...}{Additional arguments passed to internal helper functions.}
}
\value{
tibble: query, hit_cid, hit_smiles, score(NA as default)
}
\description{
Similarity search (Tanimoto) on PubChem
}
\examples{
\dontrun{
sim_hits <- compound_similarity(
query = "996",
from = "cid",
compute_score = TRUE)
}
}
