% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_BN}
\alias{compute_BN}
\title{Calculate BottleNeck (BN) score for PPI network}
\usage{
compute_BN(g)
}
\arguments{
\item{g}{An \code{igraph} object (typically an undirected, unweighted PPI network).}
}
\value{
The input graph \code{g} with an additional vertex attribute
\code{BN} storing the BottleNeck score for each node.
}
\description{
BottleNeck (BN) identifies vertices that frequently serve as bottlenecks
in shortest-path trees rooted at all vertices, following the definition
in cytoHubba (Chin et al., 2014).
}
\references{
Chin, C.H., Chen, S.H., Wu, H.H., Ho, C.W., Ko, M.T., & Lin, C.Y. (2014).
cytoHubba: identifying hub objects and sub-networks from complex interactome.
\emph{BMC Systems Biology}, 8(Suppl 4), S11.
}
\keyword{internal}
