% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_DMNC}
\alias{compute_DMNC}
\title{calculate DMNC (Density of Maximum Neighborhood Component) of PPI network}
\usage{
compute_DMNC(g, alpha = 1.7)
}
\arguments{
\item{g}{An igraph object of PPI network}

\item{alpha}{Numeric exponent used to penalize large neighborhood components. The original cytoHubba implementation uses \code{alpha = 1.7}.}
}
\value{
An igraph object containing DMNC score
}
\description{
calculate DMNC (Density of Maximum Neighborhood Component) of PPI network
}
\references{
Chin, C.H., Chen, S.H., Wu, H.H., Ho, C.W., Ko, M.T., & Lin, C.Y. (2014).
cytoHubba: identifying hub objects and sub-networks from complex interactome.
\emph{BMC Systems Biology}, 8(Suppl 4), S11.
https://doi.org/10.1186/1752-0509-8-S4-S11
}
\keyword{internal}
