% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_MCC}
\alias{compute_MCC}
\title{compute MCC (maximum clique centrality) of PPI network.}
\usage{
compute_MCC(graph)
}
\arguments{
\item{graph}{A igraph object from PPI network.}
}
\value{
An igraph object containing MCC.
}
\description{
compute MCC (maximum clique centrality) of PPI network.
}
