% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_MNC}
\alias{compute_MNC}
\title{calculate MNC (Maximum Neighborhood Component) of PPI network}
\usage{
compute_MNC(g)
}
\arguments{
\item{g}{An igraph object of PPI network}
}
\value{
An igraph object containing MNC score
}
\description{
calculate MNC (Maximum Neighborhood Component) of PPI network
}
\references{
Chin, C.H., Chen, S.H., Wu, H.H., Ho, C.W., Ko, M.T., & Lin, C.Y. (2014).
cytoHubba: identifying hub objects and sub-networks from complex interactome.
\emph{BMC Systems Biology}, 8(Suppl 4), S11.
https://doi.org/10.1186/1752-0509-8-S4-S11
}
\keyword{internal}
