% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_Stress}
\alias{compute_Stress}
\title{Calculate Stress Centrality for PPI network}
\usage{
compute_Stress(g)
}
\arguments{
\item{g}{An \code{igraph} object.}
}
\value{
The input \code{igraph} object with a new vertex attribute \code{Stress}.
}
\description{
This function calculates the Stress Centrality for each vertex in the graph.
Stress centrality of a node v is defined as the number of shortest paths
passing through v.
}
\details{
The Stress Centrality \eqn{C_{str}(v)} is calculated as:
\deqn{C_{str}(v) = \sum_{s \neq v \neq t} \sigma_{st}(v)}
where \eqn{\sigma_{st}(v)} is the number of shortest paths from node \eqn{s}
to node \eqn{t} which use the node \eqn{v}.

This implementation uses a variation of Brandes' algorithm. Note that for
large graphs, this computation can be time-consuming as it essentially
involves all-pairs shortest path calculations with complexity \eqn{O(VE)}.
}
\references{
Chin, C.H., Chen, S.H., Wu, H.H., Ho, C.W., Ko, M.T., & Lin, C.Y. (2014).
cytoHubba: identifying hub objects and sub-networks from complex interactome.
\emph{BMC Systems Biology}, 8(Suppl 4), S11.
}
\keyword{internal}
