% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DN_deg}
\alias{DN_deg}
\title{Differentially expressed genes (DEGs) in diabetic nephropathy}
\format{
A data frame with 4 columns:
\describe{
\item{GeneID}{Ensembl gene identifier}
\item{padj}{adjusted p-value (FDR)}
\item{log2FoldChange}{log2 fold change (advanced DN vs control)}
\item{Symbol}{gene symbol}
}
}
\source{
GEO dataset \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE142025}{GSE142025}
}
\usage{
DN_deg
}
\description{
A data frame of significantly differentially expressed genes from
GSE142025, comparing advanced diabetic nephropathy (DN) with control samples.
}
\details{
DEGs were filtered using padj < 0.05 and |log2FC| > 1.5.
}
\examples{
data(DN_deg)
head(DN_deg)
}
\keyword{datasets}
