% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dn_gcds}
\alias{dn_gcds}
\title{Diabetic Nephropathy (DN) Associated Genes from GeneCards}
\format{
A character vector with 4760 elements:
\describe{
\item{dn_gcds}{A list of HGNC gene symbols representing potential targets for DN.}
}
}
\source{
\url{https://www.genecards.org/}
}
\usage{
data(dn_gcds)
}
\description{
A dataset containing gene symbols associated with Diabetic Nephropathy (DN),
retrieved from the GeneCards database. All targets in this list have a
Relevance Score greater than 0.
}
\examples{
# Load the data
data(dn_gcds)

# Check the first few genes
head(dn_gcds)

}
\keyword{datasets}
