% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dn_targets}
\alias{dn_targets}
\title{Diabetic Nephropathy (DN) Associated Genes}
\format{
A character vector with 4760 elements:
\describe{
\item{dn_targets}{A list of HGNC gene symbols representing potential targets for DN.}
}
}
\source{
\url{https://www.genecards.org/}
}
\usage{
data(dn_targets)
}
\description{
A dataset containing gene symbols associated with Diabetic Nephropathy (DN),
retrieved from the GeneCards database. All targets in this list have an
inference score greater than 0.
}
\examples{
# Load the data
data(dn_targets)

# Check the first few genes
head(dn_targets)

}
\keyword{datasets}
