% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCODE.R
\name{.mcode_find_clusters}
\alias{.mcode_find_clusters}
\title{Molecular complex prediction (step 2 for MCODE method)}
\usage{
.mcode_find_clusters(g, scores, vwp, max_depth)
}
\arguments{
\item{g}{An igraph object (PPI networks).}

\item{scores}{A numeric vector containing each vertex's score after running \code{.mcode_score_nodes()}.}

\item{vwp}{Numeric. Node Score Cutoff (Vertex Weight Percentage). Default 0.2.}

\item{max_depth}{Numeric. Maximum recursion depth for cluster finding (to prevent stack overflow on huge networks). Default 100.}
}
\value{
A list containing clusters.
}
\description{
Molecular complex prediction (step 2 for MCODE method)
}
\keyword{internal}
