% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCODE.R
\name{.mcode_post_process}
\alias{.mcode_post_process}
\title{Post-processing clusters (step 3 for MCODE method)}
\usage{
.mcode_post_process(g, clusters, scores, haircut, fluff, fdt, k_core_threshold)
}
\arguments{
\item{g}{An igraph object (PPI networks)}

\item{clusters}{A list containing cluster results after running \code{.mcode_find_clusters()}.}

\item{scores}{A numeric vector containing each vertex's score after running \code{.mcode_score_nodes()}.}

\item{haircut}{Logical. Whether to prune singly-connected nodes. Default TRUE.}

\item{fluff}{Logical. Whether to expand cluster by adding dense neighbors. Default FALSE.}

\item{fdt}{Numeric. Fluff Node Density Cutoff. Used if fluff=TRUE. Default 0.1.}

\item{k_core_threshold}{Numeric. Filters out clusters that do not contain a k-core of at least this level. Default 2.}
}
\value{
A list containing final clusters after post-processing.
}
\description{
Post-processing clusters (step 3 for MCODE method)
}
\keyword{internal}
