% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCODE.R
\name{.mcode_score_nodes}
\alias{.mcode_score_nodes}
\title{Vertix weighting (step 1 for MCODE method)}
\usage{
.mcode_score_nodes(g, degree_cutoff)
}
\arguments{
\item{g}{An igraph object (PPi networks).}

\item{degree_cutoff}{Numeric. Nodes with degree < cutoff will get a score of 0. Default 2.}
}
\value{
A numeric vectors containing each vertex's score.
}
\description{
Vertix weighting (step 1 for MCODE method)
}
\keyword{internal}
