% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdock.R
\name{get_order}
\alias{get_order}
\title{Retrieve ordering function for reordering factors}
\usage{
get_order(order)
}
\arguments{
\item{order}{Ordering option. Can be one of:}
}
\value{
A function object or \code{NULL} if no ordering is requested.
}
\description{
Retrieve ordering function for reordering factors
}
\keyword{internal}
