% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_ora.R
\name{get_top_herb}
\alias{get_top_herb}
\title{Get top enriched herbs}
\usage{
get_top_herb(ora_res, top = 10)
}
\arguments{
\item{ora_res}{A tibble/data.frame from herb_ora(), must contain columns: herb, pvalue, padj.}

\item{top}{Integer, number of top herbs to keep after sorting by FDR. Default 10.}
}
\description{
Get top enriched herbs
}
\keyword{internal}
