% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_sankey_plot.R
\name{getenrichres}
\alias{getenrichres}
\title{Extract enrichment table for Sankey-Dot plotting}
\usage{
getenrichres(x, n = 10)
}
\arguments{
\item{x}{enrichResult object or data.frame containing enrichment results.}

\item{n}{Integer. The number of pathways used for visualization.}
}
\value{
a tibble with necessary columns for sankey-dot plot.
}
\description{
Extract enrichment table for Sankey-Dot plotting
}
\keyword{internal}
