% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocircle_plot.R
\name{getGores}
\alias{getGores}
\title{Prepare input data for enrich_circle_plot}
\usage{
getGores(x, up_genes = NULL, down_genes = NULL, top = 10)
}
\arguments{
\item{x}{enrichResult object or data.frame from clusterProfiler
with columns: ID, ONTOLOGY, BgRatio, p.adjust, geneID, Count, GeneRatio and RichFactor.}

\item{up_genes}{Character vector of up-regulated genes (same ID space as \code{geneID}).}

\item{down_genes}{Character vector of down-regulated genes (same ID space as \code{geneID}).}

\item{top}{Numeric. Number of pathways to keep per category (by ascending \code{p.adjust}). Default 10.}
}
\description{
Prepare input data for enrich_circle_plot
}
