% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCODE.R
\name{getMCODE_res}
\alias{getMCODE_res}
\title{Extract MCODE analysis results as a data frame}
\usage{
getMCODE_res(g, only_clusters = FALSE)
}
\arguments{
\item{g}{An igraph object processed by \code{runMCODE}.}

\item{only_clusters}{Logical. If TRUE, returns only nodes that belong to a cluster. Default FALSE.}
}
\value{
A data frame with columns:
\item{name}{Gene Symbol / Node Name}
\item{mcode_score}{Node score (k-core * density)}
\item{cluster}{Cluster label (Module_X)}
\item{module_score}{Score of the module (Density * Size)}
\item{is_seed}{Logical. Whether the node is the seed of the module}
}
\description{
Extract MCODE analysis results as a data frame
}
