% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_ora.R
\name{herb_ora}
\alias{herb_ora}
\title{Herb-target ORA (multiple herbs) with TCM targets}
\usage{
herb_ora(
  genes,
  herb_sets = NULL,
  type = c("Herb_cn_name", "Herb_pinyin_name", "Herb_en_name"),
  universe = NULL,
  p_adjust_method = "BH"
)
}
\arguments{
\item{genes}{Character vector of query genes.}

\item{herb_sets}{A vector of herbs' names.}

\item{universe}{Optional background size or background gene vector.
If \code{NULL}, defaults to the union of \code{tcmdata$target}.}

\item{p_adjust_method}{Multiple-testing correction method (default "BH").}
}
\value{
A tibble with columns: herb, ngene, setsize, indrg, pvalue, padj, overlap
}
\description{
Herb-target ORA (multiple herbs) with TCM targets
}
