% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_ora.R
\name{herb_ora_one}
\alias{herb_ora_one}
\title{Herb-target ORA (single herb) with TCM targets}
\usage{
herb_ora_one(genes, herb_targets, universe = NULL)
}
\arguments{
\item{genes}{Character vector of query genes.}

\item{herb_targets}{Character vector of targets for a single herb.}

\item{universe}{Optional. Either an integer background size or a character vector
of background genes. If \code{NULL}, it defaults to the union of all targets in \code{tcmdata$target}.}
}
\value{
List with fields: ngene, setsize, indrg, pvalue, overlap
}
\description{
Herb-target ORA (single herb) with TCM targets
}
