% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_sankey_plot.R
\name{insert_spacer_nodes}
\alias{insert_spacer_nodes}
\title{Insert spacer nodes between strata for better visual separation in Sankey plots}
\usage{
insert_spacer_nodes(dat)
}
\arguments{
\item{dat}{A data frame or tibble.}
}
\value{
A tibble with additional "spacer" rows and updated factor levels
in the \code{stratum} column.
}
\description{
Insert spacer nodes between strata for better visual separation in Sankey plots
}
\keyword{internal}
