% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_sankey_plot.R
\name{make_colors}
\alias{make_colors}
\title{set colors for sankey and dot plot}
\usage{
make_colors(items, colors, insert = NULL)
}
\arguments{
\item{items}{Character vector of unique item names to assign colors to.}

\item{colors}{Character vector of base colors to use for mapping.}

\item{insert}{Optional character vector of additional colors to insert into the palette.}
}
\description{
set colors for sankey and dot plot
}
\keyword{internal}
