% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{norm01}
\alias{norm01}
\title{Normalize a numeric vector to the range \eqn{[0, 1]}.}
\usage{
norm01(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na_rm}{Logical; whether to ignore non-finite values when computing
the range. Defaults to \code{TRUE}.}
}
\value{
Numeric vector which has been scaled.
}
\description{
A simple and robust 0–1 scaler. Handles NA, Inf, or constant values by
returning 0.5 for those entries.
}
\keyword{internal}
