% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{plot.tcm_pubmed}
\alias{plot.tcm_pubmed}
\title{Plot PubMed Search Results}
\usage{
\method{plot}{tcm_pubmed}(
  x,
  type = c("trend", "journal"),
  N = 10,
  bar_col = "#4E79A7",
  line_col = "#2F4858",
  point_col = "#B85C5C",
  ...
)
}
\arguments{
\item{x}{An object of class 'tcm_pubmed'.}

\item{type}{Character. Type of plot to return: "trend", or "journal". Default is "trend".}

\item{N}{Integer. The number of top journals to display in the bar chart. Defaults to 10.}

\item{bar_col}{Character. Hex code or name for the fill color of the bars in the journal chart. Defaults to "#3498db".}

\item{line_col}{Character. Hex code or name for the color of the trend line. Defaults to "#2c3e50".}

\item{point_col}{Character. Hex code or name for the color of the points on the trend line. Defaults to "#e74c3c".}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A ggplot object or a list of plots.
}
\description{
Visualizes the results of a PubMed search, showing the annual publication trend or the top journals.
}
\examples{
\dontrun{
## get result from pubmed
result <- get_pubmed_data(tcm_name = "Artemisinin", 
disease_name = "Malaria", 
year_range = c(2020, 2025),
retmax = 100,
email = "your email")

## plot
p1 <- plot(result, type = "trend")
p1

p2 <- plot(result, type = "journal")
p2

aplot::plot_list(p1, p2)
}

}
