% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plots.R
\name{plot_enet_coefs}
\alias{plot_enet_coefs}
\title{Bar chart of LASSO / Elastic Net / Ridge coefficients}
\usage{
plot_enet_coefs(result, top_n = 20)
}
\arguments{
\item{result}{A \code{tcm_ml} object from \code{\link[=ml_enet]{ml_enet()}}, \code{\link[=ml_lasso]{ml_lasso()}} or
\code{\link[=ml_ridge]{ml_ridge()}}.}

\item{top_n}{Number of top features to show. Default 20.}
}
\value{
A \code{ggplot} object.
}
\description{
Horizontal bar chart coloured by sign (positive / negative).
}
