% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plots.R
\name{plot_enet_cv}
\alias{plot_enet_cv}
\title{Plot Elastic Net / LASSO / Ridge CV curve}
\usage{
plot_enet_cv(result)
}
\arguments{
\item{result}{A \code{tcm_ml} object from \code{\link[=ml_enet]{ml_enet()}}, \code{\link[=ml_lasso]{ml_lasso()}} or
\code{\link[=ml_ridge]{ml_ridge()}}.}
}
\value{
A base R plot (invisible \code{NULL}).
}
\description{
Plot Elastic Net / LASSO / Ridge CV curve
}
