% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plots.R
\name{plot_rf_importance}
\alias{plot_rf_importance}
\title{Plot RF variable importance}
\usage{
plot_rf_importance(result, top_n = 20)
}
\arguments{
\item{result}{A \code{tcm_ml} object from \code{\link[=ml_rf]{ml_rf()}}.}

\item{top_n}{Integer; number of top features to show.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot RF variable importance
}
