% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plots.R
\name{plot_svm_rfe_curve}
\alias{plot_svm_rfe_curve}
\title{Plot SVM-RFE accuracy profile}
\usage{
plot_svm_rfe_curve(result)
}
\arguments{
\item{result}{A \code{tcm_ml} object from \code{\link[=ml_svm_rfe]{ml_svm_rfe()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot SVM-RFE accuracy profile
}
