% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{plot_wgcna_modules}
\alias{plot_wgcna_modules}
\title{Plot WGCNA modules}
\usage{
plot_wgcna_modules(
  x,
  block = 1L,
  linewidth = 0.25,
  base_size = 7,
  base_family = "Arial"
)
}
\arguments{
\item{x}{A \code{tcm_wgcna} object.}

\item{block}{Block number to plot.}

\item{linewidth}{Dendrogram line width.}

\item{base_size, base_family}{Text size and font family.}
}
\value{
A ggplot-compatible object.
}
\description{
Plot WGCNA modules
}
