% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{plot_wgcna_traits}
\alias{plot_wgcna_traits}
\title{Plot WGCNA module-trait associations}
\usage{
plot_wgcna_traits(
  x,
  show_labels = TRUE,
  digits = 2,
  label_size = 2.4,
  base_size = 7,
  base_family = "Arial"
)
}
\arguments{
\item{x}{A \code{tcm_wgcna} object containing trait associations.}

\item{show_labels}{Show correlations and P values in cells.}

\item{digits}{Number of correlation digits.}

\item{label_size}{Label size.}

\item{base_size, base_family}{Text size and font family.}
}
\value{
A ggplot object.
}
\description{
Plot WGCNA module-trait associations
}
