% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNodeHeatmap.R
\name{PlotNodeHeatmap}
\alias{PlotNodeHeatmap}
\title{Plot PPI nodes metrics Heatmap}
\usage{
PlotNodeHeatmap(
  data,
  id_col = "name",
  select_cols = NULL,
  colors = c("#2166AC", "white", "#B2182B"),
  cluster_rows = TRUE,
  cluster_cols = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the nodes names and metric values.}

\item{id_col}{Character. The column name serving as row names. Default is "name".}

\item{select_cols}{Character Vector. The specific columns to include in the heatmap. Default all columns except \code{id_col} are used.}

\item{colors}{Vector of length 3. Colors for low (-2), zero (0), and high (2) values. Default is c("#2166AC", "white", "#B2182B").}

\item{cluster_rows}{Logical. Whether to perform hierarchical clustering on rows. Default is TRUE.}

\item{cluster_cols}{Logical. Whether to perform hierarchical clustering on columns. Default is FALSE.}

\item{...}{Additional arguments passed to \code{ComplexHeatmap::Heatmap}.}
}
\value{
A Heatmap object.
}
\description{
Plot PPI nodes metrics Heatmap
}
\examples{
data(demo_ppi)
ppi <- compute_nodeinfo(demo_ppi)
rk_res <- rank_ppi_nodes(ppi)[[2]]
selected_cols <- colnames(rk_res)[c(2, 3, 4, 6, 9, 12, 14, 15, 16)]
p1 <- PlotNodeHeatmap(rk_res, select_cols = selected_cols)
print(p1)
}
