% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_task_agent.R
\name{print.tcm_workflow}
\alias{print.tcm_workflow}
\title{Print method for TCM workflow}
\usage{
\method{print}{tcm_workflow}(x, ...)
}
\arguments{
\item{x}{A tcm_workflow object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for TCM workflow
}
