% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_cluster.R
\name{run_fastgreedy}
\alias{run_fastgreedy}
\title{Compute Fast Greedy Clustering for PPI Network}
\usage{
run_fastgreedy(g, weights = NULL)
}
\arguments{
\item{g}{An \code{igraph} object.}

\item{weights}{Numeric vector, character, NULL, or NA. Edge weights to use for
clustering. If NULL (default), the function attempts to use the \code{weight}
edge attribute first, then the \code{score} edge attribute. If a character
value is supplied, it is treated as the edge attribute name. Set to NA to
perform unweighted clustering.}
}
\value{
The graph with \code{fastgreedy_cluster} and plot-ready \code{cluster}
attributes.
}
\description{
This function performs fast greedy modularity optimization on an igraph object.
It is suitable for detecting communities in undirected PPI networks.
}
\examples{
data(demo_ppi)
library(igraph)
ppi <- run_fastgreedy(demo_ppi)
head(V(ppi)$fastgreedy_cluster)

}
