% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_herb}
\alias{search_herb}
\title{Retrieve Herbs and Retrieve Associated Molecules and Targets}
\usage{
search_herb(herb, type)
}
\arguments{
\item{herb}{A character vector containing the names of herbs to be queried.}

\item{type}{A string indicating the type of herb names provided. Must be one of \code{"Herb_cn_name"},
\code{"Herb_pinyin_name"}, or \code{"Herb_en_name"}.}
}
\value{
A \code{data.frame} with three columns: \code{herb}, \code{molecule},
and \code{target}, containing the corresponding information for the specified herbs.
Rows with \code{NA} values are excluded.
}
\description{
This function retrieves herb, molecule, and target information from the internal dataset, \code{tcm_data},
based on specified herb names and their corresponding name types (Chinese, Pinyin, or English).
}
\examples{
\dontrun{
# Example usage with Chinese herb names
herbs <- c("灵芝")
lz <- search_herb(herb = herbs, type = "Herb_cn_name")
print(lz)

# Example usage with pinyin herb names
herbs <- c("Ginseng", "Licorice")
comp <- search_herb(herb = herbs, type = "Herb_pinyin_name")
print(comp)
}

}
