% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_utils.R
\name{similarity_search}
\alias{similarity_search}
\title{Similarity search (Tanimoto) on PubChem}
\usage{
similarity_search(
  query,
  from = c("smiles", "cid", "inchikey", "name"),
  threshold = 90,
  topn = 10,
  ...
)
}
\arguments{
\item{query}{identifier (CID/SMILES/InChIKey/Name)}

\item{from}{one of c("cid","smiles","inchikey","name")}

\item{threshold}{integer 0–100 (percent), e.g. 90}

\item{topn}{max records to return. Default 10.}

\item{...}{Additional arguments passed to internal helper functions.}
}
\value{
tibble: query, hit_cid, score
}
\description{
Similarity search (Tanimoto) on PubChem
}
