% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCM_alluvial.R
\name{TCM_alluvial}
\alias{TCM_alluvial}
\title{Draw an Alluvial Diagram for Hierarchical Data}
\usage{
TCM_alluvial(
  data,
  text.size = 3,
  text.position = 1,
  x.axis.text.size = 12,
  flow.alpha = 0.5,
  palette = "rainbow_wh_rd",
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with at least two columns representing hierarchical levels.}

\item{text.size}{Numeric. Font size of the node labels. Default is 3.}

\item{text.position}{Numeric. Horizontal justification of the node labels. Typically 0 (left), 0.5 (center), or 1 (right). Default is 1.}

\item{x.axis.text.size}{Numeric. Font size of the x-axis text labels. Default is 12.}

\item{flow.alpha}{Numeric (between 0 and 1). Transparency of the flow paths. Default is 0.5.}

\item{palette}{Character. Name of the color palette to use from \code{cols4all::c4a}. Default is \code{"rainbow_wh_rd"}.}
}
\value{
A \code{ggplot} object representing the alluvial diagram.
}
\description{
Draw an Alluvial Diagram for Hierarchical Data
}
