% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{compute_MCC}
\alias{compute_MCC}
\title{compute MCC (maximum clique centrality) of PPI network.}
\usage{
compute_MCC(graph)
}
\arguments{
\item{graph}{A igraph object from PPI network.}
}
\value{
An igraph object containing MCC
}
\description{
compute MCC (maximum clique centrality) of PPI network.
}
\references{
Chin, C.H., Chen, S.H., Wu, H.H., Ho, C.W., Ko, M.T., & Lin, C.Y. (2014).
cytoHubba: identifying hub objects and sub-networks from complex interactome.
\emph{BMC Systems Biology}, 8(Suppl 4), S11.
https://doi.org/10.1186/1752-0509-8-S4-S11
}
\keyword{internal}
