% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_utils.R
\name{getprops}
\alias{getprops}
\title{get properties for compounds from PubChem}
\usage{
getprops(
  cid,
  properties = c("MolecularFormula", "MolecularWeight", "IUPACName", "CanonicalSMILES",
    "InChIKey", "XLogP"),
  ...
)
}
\arguments{
\item{cid}{Integer/numeric/character vector of PubChem CIDs.}

\item{properties}{Character vector of PubChem property keys to request.}

\item{...}{Additional arguments passed to internal helper functions.}
}
\description{
get properties for compounds from PubChem
}
\examples{
\dontrun{
herbs <- c("灵芝")
lz <- search_herb(herb = herbs, type = "Herb_cn_name")
lz_mol <- sample(unique(lz$molecule), 5, replace = FALSE)
lz_mol_cid <- resolve_cid(lz_mol, from = "name")
props <- getprops(lz_mol_cid)
print(props)
}
}
