% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvenndata.R
\name{getvenndata}
\alias{getvenndata}
\title{Construct a Logical Matrix for Venn Diagram Visualization}
\usage{
getvenndata(..., set_names = NULL)
}
\arguments{
\item{...}{2 to 4 character vectors representing different gene sets or element sets.}

\item{set_names}{Optional character vector to assign custom names to the sets.
If not provided, names will be automatically assigned as "Set1", "Set2", etc.}
}
\value{
A \code{data.frame} with:
\itemize{
\item \code{Element}: all unique elements across the sets.
\item One logical column per input set indicating presence (TRUE/FALSE) of each element.
}
}
\description{
This function takes 2 to 4 named character vectors as input, removes duplicates,
and returns a data frame in logical format that is compatible with \code{ggvenn} for Venn diagram plotting.
}
