% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdock.R
\name{ggdock}
\alias{ggdock}
\title{Plot molecule–target docking affinity heatmaps}
\usage{
ggdock(
  dock_data,
  order = NULL,
  type = "dot",
  point_size = 8,
  base_size = 12,
  angle = 50,
  hjust = 1,
  vjust = 1,
  palette = "ggthemes::Orange-Blue-White Diverging",
  label = FALSE,
  label_digits = 2,
  label_size = 3,
  label_color = "black",
  label_family = "sans",
  label_fontface = "plain",
  ...
)
}
\arguments{
\item{dock_data}{A numeric matrix or data frame.}

\item{order}{Reordering method for factor levels (optional).}

\item{type}{Type of visualization, either \code{"dot"} or \code{"tile"}.}

\item{point_size}{Numeric. Point size used in dot plots. Default is 6.}

\item{base_size}{Numeric. Base font size for the theme. Default is 12.}

\item{angle}{Numeric. Rotation angle for x-axis text labels. Default is 50.}

\item{hjust, vjust}{Numeric. Horizontal and vertical justification for x-axis labels.}

\item{palette}{Character. Name of the continuous color palette to use. Default is \code{"ggthemes::Orange-Blue-White Diverging"}.}

\item{label}{Logical. If \code{TRUE}, print affinity numbers on the plot. Default \code{FALSE}.}

\item{label_digits}{Integer. Number of digits to show for the affinity text. Default \code{2}.}

\item{label_size}{Numeric. Text size for affinity labels. Default \code{3.5}.}

\item{label_color}{Character. Text color for affinity labels. Default \code{"black"}.}

\item{label_family}{Character. Font family for labels (e.g., \code{"sans"}, \code{"Times"}). Default \code{"sans"}.}

\item{label_fontface}{Character. Font face for labels: \code{"plain"}, \code{"bold"}, \code{"italic"}, \code{"bold.italic"}. Default \code{"plain"}.}

\item{...}{Additional parameters passed to \code{geom_point()} or \code{geom_tile()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot molecule–target docking affinity heatmaps
}
