% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_sankey_plot.R
\name{ggdot_sankey}
\alias{ggdot_sankey}
\title{Dot–Sankey plot for enrichment results}
\usage{
ggdot_sankey(
  enrich_obj,
  n = 10,
  axis_order = c("Pathway", "Gene"),
  id_colors = c("#8DD3C7", "#FFFFB3", "#BEBADA", "#FB8072", "#80B1D3", "#FDB462",
    "#B3DE69", "#FCCDE5", "#D9D9D9", "#BC80BD", "#CCEBC5", "#FFED6F"),
  desc_colors = c("#E64B35", "#4DBBD5", "#00A087", "#3C5488", "#F39B7F", "#8491B4",
    "#91D1C2", "#C8DC96", "#7E6148", "#B09C85", "#6A5ACD", "#A0522D"),
  id_y_pos = 1.1,
  desc_y_pos = 1,
  pathway_wrap = 50,
  sankey_text_size = 4,
  bubble_size_range = c(3, 8),
  dot_palette = "RdBu",
  dot_x_var = c("GeneRatio", "RichFactor", "FoldEnrichment"),
  bubble_p_label = "p.adjust",
  sankey_width = 2,
  dot_width = 1,
  font_family = "sans",
  font_face = "bold",
  sankey_lab = "Gene-Pathway",
  seed = 2025,
  ...
)
}
\arguments{
\item{enrich_obj}{enrichResult object or data.frame containing enrichment results.}

\item{n}{Integer. The number of pathways to visualize (top \code{n} ranked by significance).}

\item{axis_order}{Character vector of length 2, specifying the order of axes in
the Sankey diagram. Default is \code{c("Pathway", "Gene")}.}

\item{id_colors}{Character vector of base colors for gene nodes.}

\item{desc_colors}{Character vector of base colors for pathway nodes.}

\item{id_y_pos}{Numeric constant controlling the y-position of “Gene” strata in the Sankey diagram. Default is \code{1.1}.}

\item{desc_y_pos}{Numeric constant controlling the y-position of “Pathway” strata in the Sankey diagram. Default is \code{1.0}.}

\item{pathway_wrap}{Integer. The maximum line width for pathway label wrapping. Default is \code{50}.}

\item{sankey_text_size}{Numeric. Font size for text labels in the Sankey diagram. Default is \code{4}.}

\item{bubble_size_range}{Numeric vector of length 2. Range of point sizes in the dot plot. Default is \code{c(3, 8)}.}

\item{dot_palette}{Character. Name of the RColorBrewer palette used for color gradients in the dot plot. Default is \code{"RdBu"}.}

\item{dot_x_var}{Character. Variable used for the x-axis in the dot plot.}

\item{bubble_p_label}{Character. The column name for significance values used to color dotsd. Default is \code{p.adjust}.}

\item{sankey_width}{Numeric. Relative width of the Sankey panel in the combined plot. Default is \code{2}.}

\item{dot_width}{Numeric. Relative width of the dot plot panel in the combined plot. Default is \code{1}.}

\item{font_family}{Character. Font family for all text elements. Default is \code{"Arial"}.}

\item{font_face}{Character. Font face for all text. Default is \code{"bold"}.}

\item{sankey_lab}{Character. Label for the x-axis of the Sankey diagram. Default is \code{"Gene-Pathway"}.}

\item{seed}{Integer. Random seed for reproducibility of layout. Default is \code{2025}.}

\item{...}{Additional arguments passed to internal helper functions.}
}
\value{
a ggplot object containing dot plot and sankey plot.
}
\description{
Dot–Sankey plot for enrichment results
}
