% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglollipop.R
\name{gglollipop}
\alias{gglollipop}
\title{Lollipop Plot for Enrichment Results}
\usage{
gglollipop(
  enrich_obj,
  x = "RichFactor",
  top_n = 10,
  orderBy = NULL,
  text.col = "black",
  text.size = 8,
  text.width = 35,
  palette = "RdBu",
  line.col = "grey60",
  line.type = "solid",
  line.size = 0.9,
  plot_title = NULL,
  show_count = TRUE,
  ...
)
}
\arguments{
\item{enrich_obj}{An enrichment result object from clusterProfiler.}

\item{x}{Character. The variable used for the x-axis. Default is \code{"RichFactor"}.}

\item{top_n}{Integer. Number of top enriched terms to display. Default is 10.}

\item{orderBy}{Character. Variable used to order the y-axis terms. Default is \code{"x"}.}

\item{text.col}{Character. Colors for text. Default is \code{black}.}

\item{text.size}{Numeric. Font size for axis text and title. Default is 8.}

\item{text.width}{Numeric. Font width for axis text and title. Default is 35.}

\item{palette}{Character. Color palette name from \code{RColorBrewer} to use for dot color. Default is \code{"RdBu"}.}

\item{line.col}{Character. Color of the segment lines. Default is \code{"grey60"}.}

\item{line.type}{Character. Line type for segments. Default is \code{"solid"}.}

\item{line.size}{Numeric. Line width for segments. Default is 0.9.}

\item{plot_title}{Character. Optional plot title. Default is \code{NULL}.}

\item{show_count}{Logical. Whether to display the count value as a text label next to each dot. Default is TRUE}

\item{...}{Additional arguments passed to internal helper functions.}
}
\value{
A \code{ggplot} object showing a lollipop-style enrichment plot.
}
\description{
Lollipop Plot for Enrichment Results
}
