% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plots.R
\name{plot_ml_venn}
\alias{plot_ml_venn}
\title{Venn diagram of selected genes}
\usage{
plot_ml_venn(ml_list)
}
\arguments{
\item{ml_list}{A \code{tcm_ml_list} from \code{\link[=run_ml_screening]{run_ml_screening()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Internally calls \code{\link[=getvenndata]{getvenndata()}} + \code{\link[=ggvenn_plot]{ggvenn_plot()}}.
}
