% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{ppi_subset}
\alias{ppi_subset}
\title{Subset PPI network by edge score and top-n node degree}
\usage{
ppi_subset(ppi_obj, n = NULL, score_cutoff = 0.7)
}
\arguments{
\item{ppi_obj}{An igraph object with edge attribute 'score'}

\item{n}{Number of top-degree nodes to keep. Default is NULL.}

\item{score_cutoff}{Minimum edge score to keep. Default is 0.7.}
}
\value{
A subgraph of the original PPI network
}
\description{
Subset PPI network by edge score and top-n node degree
}
