% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_herb_graph.R
\name{prepare_herb_graph}
\alias{prepare_herb_graph}
\title{Prepare an igraph object from a Herb–Molecule–Target data frame}
\usage{
prepare_herb_graph(
  df,
  n = NULL,
  herb_col = "herb",
  molecule_col = "molecule",
  target_col = "target",
  compute_metrics = TRUE
)
}
\arguments{
\item{df}{A \code{data.frame} with three columns representing herb–molecule–target relationships.}

\item{n}{Integer. Optional. The number of rows to use from df. Default is 'nrow(df)'}

\item{herb_col}{Character. The column name corresponding to herbs. Default is \code{"herb"}.}

\item{molecule_col}{Character. The column name corresponding to molecules/compounds. Default is \code{"molecule"}.}

\item{target_col}{Character. The column name corresponding to targets. Default is \code{"target"}.}

\item{compute_metrics}{Logical. Whether to compute and attach node-level metrics. Default is TRUE.}
}
\value{
A directed \code{igraph}
}
\description{
Prepare an igraph object from a Herb–Molecule–Target data frame
}
