% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_utils.R
\name{rank_ppi_nodes}
\alias{rank_ppi_nodes}
\title{Rank PPI nodes by integrated network centrality}
\usage{
rank_ppi_nodes(
  g,
  metrics = c("degree", "betweenness", "closeness", "eccentricity", "radiality",
    "Stress", "MCC", "MNC", "DMNC", "BN", "EPC"),
  weights = NULL,
  use_weight = TRUE,
  na_rm = TRUE
)
}
\arguments{
\item{g}{An igraph object that has already been processed by \code{compute_nodeinfo()} or have added node metrics manually.}

\item{metrics}{Character vector; which vertex attributes to use for scoring. Defaults are the same as Cytohubba.}

\item{weights}{Numeric vector of the same length as metrics; relative weights for each metric. If NULL (default), all metrics are equally weighted.}

\item{use_weight}{Logical; whether use weighted metrics(beweenness and closeness) instead. Default is TRUE.}

\item{na_rm}{Logical; if TRUE, NAs are ignored in normalization (set to 0.5).}
}
\value{
A list with:
\item{graph}{igraph object with added vertex attributes \code{Score_network} and \code{Rank_network}.}
\item{table}{data.frame with node-level metrics and scores, sorted by \code{Rank_network}.}
}
\description{
Rank PPI nodes by integrated network centrality
}
\examples{
data(demo_ppi)
library(igraph)
ppi <- compute_nodeinfo(demo_ppi)
rank_res <- rank_ppi_nodes(ppi)

# update igraph object and extract rank info
ppi <- rank_res[[1]]
rank_df <- rank_res[[2]]
print(rank_df)

}
