% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCM_sankey.R
\name{TCM_sankey}
\alias{TCM_sankey}
\title{Create a Sankey Diagram for Hierarchical TCM or Omics Data}
\usage{
TCM_sankey(
  data,
  text.size = 3,
  text.position = 1,
  x.axis.text.size = 12,
  flow.alpha = 0.5,
  palette = "rainbow_wh_rd",
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} for sankey_plot.}

\item{text.size}{Numeric. Size of node label text. Default is 3.}

\item{text.position}{Numeric. Horizontal justification of node labels; typically 0 (left),
0.5 (center), or 1 (right). Default is 1.}

\item{x.axis.text.size}{Numeric. Font size of x-axis labels. Default is 12.}

\item{flow.alpha}{Numeric (0–1). Transparency of the Sankey flows. Default is 0.5.}

\item{palette}{Character. Name of the color palette to use from \code{cols4all::c4a_pals()}.
Default is \code{"rainbow_wh_rd"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a Sankey Diagram for Hierarchical TCM or Omics Data
}
