% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCM_sankey.R
\name{TCM_sankey}
\alias{TCM_sankey}
\title{TCM Sankey Plot for herb-molecule-target}
\usage{
TCM_sankey(
  data,
  axis_order = c("herb", "molecule", "target"),
  herb_cols = c("#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd", "#8c564b",
    "#e377c2", "#7f7f7f"),
  mol_cols = c("#bcbd22", "#17becf", "#aec7e8", "#ffbb78", "#98df8a", "#ff9896",
    "#c5b0d5", "#E41A1C", "#377EB8"),
  target_cols = c("#c49c94", "#f7b6d2", "#dbdb8d", "#c7e9c0", "#f4cae4", "#e6f598",
    "#ffeda0", "#BB5234", "#BB7813", "#FF6158"),
  plot_font = "sans",
  font_face = "bold",
  font_size = 3.6,
  width = 0.05,
  alpha = 0.3,
  knot.pos = 0.3
)
}
\arguments{
\item{data}{A data frame containing at least three columns: \code{herb}, \code{molecule}, and \code{target}.}

\item{axis_order}{Character vector specifying the order of axes in the Sankey diagram. Default is \code{c("herb", "molecule", "target")}.}

\item{herb_cols}{Character vector defining the base color palette for the herb layer.}

\item{mol_cols}{Character vector defining the base color palette for the molecule.}

\item{target_cols}{Character vector defining the base color palette for the target.}

\item{plot_font}{Character string specifying the font family used for text. Default is \code{"sans"}.}

\item{font_face}{Character string specifying the font face. Default is \code{"bold"}.}

\item{font_size}{Numeric value controlling the size of node labels. Default is \code{3.5}.}

\item{width}{Numeric value controlling the width of both nodes and flows. Default is \code{0.05}.}

\item{alpha}{Numeric value controlling the transparency of the flows. Default is \code{0.3}.}

\item{knot.pos}{Numeric value (between 0 and 1) determining the curvature position of flow lines. Default is \code{0.3}.}
}
\value{
A \code{ggplot} object representing the herb–compound–target Sankey diagram.
}
\description{
TCM Sankey Plot for herb-molecule-target
}
